A naive implementation of raw data recording mode (#6)
This is an optional feature that should be off by default: storing raw
GPS data with detailed tempstamp. It is designed for advanced user or
debugging. It stores data in a simple csv format and will be using a new
file every time the app starts.

The implementation is extremely naive and should be refined later.