Checkin `flutter-rust-bridge` codegen codes. (#22)
We skip running the frb codegen if the codegen is not installed. This is
mostly fine because we checked-in generated code in git. Personally I
don't like this idea, but by doing this we:
1. People that worked on rust only don't need to install flutter + the
codegen.
2. Make rust only github action faster.
3. Avoid certain issues: mostly the versioning story of
`flutter_rust_bridge_codegen` is a little weird. And there is one more
issue with the rust cache on github action that I don't fully
understand.

On the other hand, the protobuf's behavior is a lot more reasonable.