Introducing journey date (#60)
Per our offline discussion, we are introducing a new concept:
`journey_date`. The min granularity of history in this app is day and
each journey is an atomic entry attach to each day using `journey_date`.
We believe this is better than using the old `end_time` approach because
it handles timezone slightly better and make the concept of `end_time`
less overloaded. Now `end_time` is just a normal metadata.

We haven't made our stability guarantee yet. This change has
incompatible changes to maindb so all installation need a clean reset.
The archive format is backward compatible.