Fix a bug in `_loadMapState` (#99)
`geolocator.Geolocator.getLastKnownPosition` may raise exception when it
does not have enough permission.
Now I think about it, falling back to
`geolocator.Geolocator.getLastKnownPosition` does not make a lot of
sense anyway. If we don't have saved map state, it is very likely that
it is the first time opening the app. In this case, we don't have
location permission anyway.