WebView Phase 1 & 2 (#141)
Replace the FRB-based rendering to WebView based.

Currently it is the early-stage of migration, the goal is to minimize
the modification to the existing codes. Since we made a lot of
unoptimized changes, the current version may see increased energy
consumption.

Know issues:

1. Many todos for improvement 
1. When complete a journey, the punk will not disappeared automatically.
(If journey stops, no gps will be available anymore in the current
implementation.
1. The main map state serialization is not implemented yet.

Feature expected in the future:

1. more integration tests
1. incremental frontend updates
1. redesign interfaces
1. performance improvement with data structure optimization.

---------

Co-authored-by: Zed(Zijun) Chen <zed@zijun.dev>