Add new API endpoint for checking journey bitmap updates
Introduced `serve_journey_bitmap_up_to_date_by_id` to allow clients to check if the bitmap has changed since the last request. Updated routing to include this new endpoint. Note that this API is not idempotent and should be used exclusively with either this or the existing bitmap provisioned API.