Refactoring using TypeScript & Bug fixes (#272)
This will be a large PR. 
I'll try not to add new features in this PR, but it does introduce a
series of refactoring during the transition.

Refactor details:
- Use ts to replace js.
- Better layer abstraction.
- Better frontend/backend communication abstraction.
- Better frontend initialization and params handling.
- Make `index.js` smaller by moving many tedious logics out.

New features:
- add `fog_density` in `ExternalParams` to help the migration of
https://github.com/MemoLanes/MemoLanes/pull/265
- add `projection` in `ExternalParams` and debug panel as well.

Bug fixes:
- Bump maplibre to the latest version, so that the black square bug is
fixed. The workaround is removed.
- In the recent release, we have a regression in the offline rendering
experience. This is due to the newly added dark black universe
background in HTML, which makes our empty map style completely dark,
rendering the tracks invisible. In this version, we add a light-color
background so that we can distinguish the Earth and the universe.
- Fix the iOS magnifier issue. (This has been fixed in proprietary
version)